+++ sqlservice/model.py
@@ -247,9 +247,12 @@ class ModelBase(object):
         return '<{0}({1})>'.format(self.__class__.__name__, values)
 
 
-    """Function and decorator that converts a normal class into a SQLAlchemy
-    declarative base class.
-def declarative_base(cls=ModelBase, metadata=None, metaclass=ModelMeta)
+def declarative_base(cls=ModelBase,
+                     metadata=None,
+                     metaclass=ModelMeta,
+                     **kargs):
+    """Function that converts a normal class into a SQLAlchemy declarative base
+    class.
 
     Args:
         cls (type): A type to use as the base for the generated declarative
@@ -263,25 +266,28 @@ def declarative_base(cls=ModelBase, metadata=None, metaclass=ModelMeta)
         metaclass (DeclarativeMeta, optional): A metaclass or ``__metaclass__``
             compatible callable to use as the meta type of the generated
             declarative base class. If not passed in, `cls.metaclass` will be
-            used if set. Defaults to ``None``.
+            used if set. Defaults to :class:`ModelMeta`.
+
+    Keyword Args:
+        All other keyword arguments are passed to
+        ``sqlalchemy.ext.declarative.declarative_base``.
 
     Returns:
         class: Declarative base class
     """
-
-    options = {'cls': cls,
-               'name': cls.__name__}
+    kargs['cls'] = cls
+    kargs.setdefault('name', cls.__name__)
 
     if hasattr(cls, '__init__'):
-        options['constructor'] = cls.__init__
+        kargs.setdefault('constructor', cls.__init__)
 
     if metadata:
-        options['metadata'] = metadata
+        kargs['metadata'] = metadata
 
     if metaclass:
-        options['metaclass'] = metaclass
+        kargs['metaclass'] = metaclass
 
-    Base = _declarative_base(**options)
+    Base = _declarative_base(**kargs)
 
     if metaclass:
         Base.metaclass = metaclass